+++ relm4-examples/examples/macro_test.rs
@@ -77,7 +77,8 @@ impl Widgets<ButtonModel, AppModel> for ButtonWidgets {
 }
 
 pub struct AppComponents {
-    button: RelmComponent<ButtonModel, AppModel>,
+    button1: RelmComponent<ButtonModel, AppModel>,
+    button2: RelmComponent<ButtonModel, AppModel>,
 }
 
 impl Components<AppModel> for AppComponents {
@@ -87,7 +88,8 @@ impl Components<AppModel> for AppComponents {
         sender: Sender<AppMsg>,
     ) -> Self {
         AppComponents {
-            button: RelmComponent::new(model, parent_widgets, sender),
+            button1: RelmComponent::new(model, parent_widgets, sender.clone()),
+            button2: RelmComponent::new(model, parent_widgets, sender),
         }
     }
 }
@@ -108,6 +110,7 @@ impl Widgets<AppModel, ()> for AppWidgets {
                 set_margin_all?: Some(5),
                 set_spacing: 5,
 
+                append: component!(components.button1.root_widget()),
                 append: inc_button = &gtk::Button {
                     set_label: "Increment",
                     connect_clicked(sender) => move |_| {
@@ -140,7 +143,7 @@ impl Widgets<AppModel, ()> for AppWidgets {
                     attach(2, 1, 1, 1) = &gtk::Label {
                         set_label: "grid test 3",
                     },
-                    attach(2, 2, 1, 1): component!(components.button.root_widget())
+                    attach(2, 2, 1, 1): component!(components.button2.root_widget())
                 }
             },
         }