+++ src/lib.rs
@@ -34,8 +34,8 @@ use once_cell::sync::OnceCell;
 
 static APP: OnceCell<Fragile<gtk::Application>> = OnceCell::new();
 
-pub use gtk::glib::Sender;
 pub use gtk;
+pub use gtk::glib::Sender;
 
 #[cfg(feature = "tokio-rt")]
 #[cfg_attr(doc, doc(cfg(feature = "tokio-rt")))]