+++ relm4-examples/examples/entry.rs
@@ -68,6 +68,7 @@ impl AppUpdate for AppModel {
     }
 }
 
+#[derive(Debug)]
 struct FactoryWidgets {
     button: gtk::Button,
 }