+++ relm4-examples/examples/to_do.rs
@@ -9,6 +9,7 @@ struct Task {
     completed: bool,
 }
 
+#[derive(Debug)]
 struct TaskWidgets {
     label: gtk::Label,
     hbox: gtk::Box,