+++ src/factory/widgets/gtk4.rs
@@ -1,29 +1,17 @@
 use gtk::glib;
 use gtk::prelude::{BoxExt, GridExt, TreeViewExt};
 
-use crate::factory::{FactoryListView, FactoryView};
-
-/// Storing information about where new widgets can be placed
-/// inside a [`gtk::Grid`].
-#[derive(Debug)]
-pub struct GridPosition {
-    /// The number of the column.
-    pub column: i32,
-    /// The number of the row.
-    pub row: i32,
-    /// The amount of columns the widget should take.
-    pub width: i32,
-    /// The amount of rows the widget should take.
-    pub height: i32,
-}
+use crate::factory::{positions::*, FactoryListView, FactoryView};
 
 impl<Widget> FactoryView<Widget> for gtk::Box
 where
     Widget: glib::IsA<gtk::Widget>,
 {
     type Position = ();
-    fn add(&self, widget: &Widget, _position: &()) {
+    type Root = Widget;
+    fn add(&self, widget: &Widget, _position: &()) -> Widget {
         self.append(widget);
+        widget.clone()
     }
 
     fn remove(&self, widget: &Widget) {
@@ -33,15 +21,17 @@ where
 
 impl<Widget> FactoryListView<Widget> for gtk::Box
 where
-    Self: FactoryView<Widget>,
+    Self: FactoryView<Widget, Root = Widget>,
     Widget: glib::IsA<gtk::Widget>,
 {
-    fn insert_after(&self, widget: &Widget, other: &Widget) {
+    fn insert_after(&self, widget: &Widget, other: &Widget) -> Widget {
         self.insert_child_after(widget, Some(other));
+        widget.clone()
     }
 
-    fn push_front(&self, widget: &Widget) {
+    fn push_front(&self, widget: &Widget) -> Widget{
         self.prepend(widget);
+        widget.clone()
     }
 }
 
@@ -50,8 +40,11 @@ where
     Widget: glib::IsA<gtk::Widget>,
 {
     type Position = ();
-    fn add(&self, widget: &Widget, _position: &()) {
+    type Root = Widget;
+
+    fn add(&self, widget: &Widget, _position: &()) -> Widget {
         self.append(widget);
+        widget.clone()
     }
 
     fn remove(&self, widget: &Widget) {
@@ -61,22 +54,27 @@ where
 
 impl<Widget> FactoryListView<Widget> for gtk::ListBox
 where
-    Self: FactoryView<Widget>,
-    Widget: gtk::prelude::ListBoxRowExt + glib::IsA<gtk::Widget>,
+    Self: FactoryView<Widget, Root = Widget>,
+    Widget: gtk::prelude::ListBoxRowExt + glib::IsA<gtk::Widget> + Clone,
 {
-    fn insert_after(&self, widget: &Widget, other: &Widget) {
+    fn insert_after(&self, widget: &Widget, other: &Widget) -> Widget {
         self.insert(widget, other.index());
+        widget.clone()
     }
 
-    fn push_front(&self, widget: &Widget) {
+    fn push_front(&self, widget: &Widget) -> Widget {
         self.prepend(widget);
+        widget.clone()
     }
 }
 
 impl FactoryView<gtk::TreeViewColumn> for gtk::TreeView {
     type Position = ();
-    fn add(&self, widget: &gtk::TreeViewColumn, _position: &()) {
+    type Root = gtk::TreeViewColumn;
+
+    fn add(&self, widget: &gtk::TreeViewColumn, _position: &()) -> gtk::TreeViewColumn {
         self.insert_column(widget, -1);
+        widget.clone()
     }
 
     fn remove(&self, widget: &gtk::TreeViewColumn) {
@@ -89,8 +87,11 @@ where
     Widget: glib::IsA<gtk::Widget>,
 {
     type Position = ();
-    fn add(&self, widget: &Widget, _position: &()) {
+    type Root = Widget;
+
+    fn add(&self, widget: &Widget, _position: &()) -> Widget {
         self.insert(widget, -1);
+        widget.clone()
     }
 
     fn remove(&self, widget: &Widget) {
@@ -100,22 +101,39 @@ where
 
 impl<Widget> FactoryListView<Widget> for gtk::FlowBox
 where
-    Self: FactoryView<Widget>,
-    Widget: gtk::prelude::FlowBoxChildExt + glib::IsA<gtk::Widget>,
+    Self: FactoryView<Widget, Root = Widget>,
+    Widget: gtk::prelude::FlowBoxChildExt + glib::IsA<gtk::Widget> + Clone,
 {
-    fn insert_after(&self, widget: &Widget, other: &Widget) {
+    fn insert_after(&self, widget: &Widget, other: &Widget) -> Widget {
         self.insert(widget, other.index());
+        widget.clone()
     }
 
-    fn push_front(&self, widget: &Widget) {
+    fn push_front(&self, widget: &Widget) -> Widget {
         self.insert(widget, 0);
+        widget.clone()
     }
 }
 
-#[derive(Debug)]
-pub struct FixedPosition {
-    pub x: f64,
-    pub y: f64,
+impl<Widget> FactoryView<Widget> for gtk::Stack
+where
+    Widget: glib::IsA<gtk::Widget>,
+{
+    type Position = StackPageInfo;
+    type Root = Widget;
+
+    fn add(&self, widget: &Widget, position: &StackPageInfo) -> Widget {
+        if let Some(title) = &position.title {
+            self.add_titled(widget, position.name.as_deref(), title);
+        } else {
+            self.add_named(widget, position.name.as_deref());
+        }
+        widget.clone()
+    }
+
+    fn remove(&self, widget: &Widget) {
+        self.remove(widget);
+    }
 }
 
 impl<Widget> FactoryView<Widget> for gtk::Fixed
@@ -123,8 +141,11 @@ where
     Widget: glib::IsA<gtk::Widget>,
 {
     type Position = FixedPosition;
-    fn add(&self, widget: &Widget, position: &FixedPosition) {
+    type Root = Widget;
+
+    fn add(&self, widget: &Widget, position: &FixedPosition) -> Widget {
         gtk::prelude::FixedExt::put(self, widget, position.x, position.y);
+        widget.clone()
     }
 
     fn remove(&self, widget: &Widget) {
@@ -137,8 +158,9 @@ where
     Widget: glib::IsA<gtk::Widget>,
 {
     type Position = GridPosition;
+    type Root = Widget;
 
-    fn add(&self, widget: &Widget, position: &GridPosition) {
+    fn add(&self, widget: &Widget, position: &GridPosition) -> Widget {
         self.attach(
             widget,
             position.column,
@@ -146,30 +168,10 @@ where
             position.width,
             position.height,
         );
+        widget.clone()
     }
 
     fn remove(&self, widget: &Widget) {
         GridExt::remove(self, widget);
-    }
-}
-
-#[cfg(feature = "libadwaita")]
-#[cfg_attr(doc, doc(cfg(feature = "libadwaita")))]
-mod adwaita {
-    use crate::factory::FactoryView;
-    use gtk::glib;
-
-    impl<Widget> FactoryView<Widget> for adw::Carousel
-    where
-        Widget: glib::IsA<gtk::Widget>,
-    {
-        type Position = ();
-        fn add(&self, widget: &Widget, _position: &Self::Position) {
-            self.append(widget);
-        }
-
-        fn remove(&self, widget: &Widget) {
-            self.remove(widget);
-        }
     }
 }