+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Added
+
++ relm4: Re-export for gtk
+
 ## 0.1.0 - 2021-09-06
 
 ### Added