+++ CHANGES.md
@@ -14,6 +14,7 @@
 + relm4: Added with_app method that allows passing an existing gtk::Appliaction to Relm4
 + relm4: Methods to access the widgets of components
 + relm4: Re-export for gtk
++ relm4-macros: Added error message when confusing `=` and `:`
 + relm4-macros: Allow usage of visibilities other than pub
 + relm4-macros: New pre_connect_components and post_connect_components for manual components code